Update pom.xml, tentative with IKVM and a jar in addition to that remove issue in the solver.py
