+++ Sudoku.ChocoSolver/solver.py
@@ -1,15 +1,5 @@
-import numpy as np
 from pycsp3 import *
 
-instance = ((0,0,0,0,9,4,0,3,0),
-        (0,0,0,5,1,0,0,0,7),
-        (0,8,9,0,0,0,0,4,0),
-        (0,0,0,0,0,0,2,0,8),
-        (0,6,0,2,0,1,0,5,0),
-        (1,0,2,0,0,0,0,0,0),
-        (0,7,0,0,0,0,5,2,0),
-        (9,0,0,0,6,5,0,0,0),
-        (0,4,0,9,7,0,0,0,0))
 
 # clues = data  # if not 0, clues[i][j] is a value imposed at row i and col j
 clues = instance
@@ -33,5 +23,5 @@ r = []
 
 if solve() is SAT:
     r = list(values(grid))
-    
-print(r)
+else:
+    print('No solution was found')
